requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
python-dotenv>=1.0.0

# AI + HTTP (existing)
anthropic>=0.25.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
